:recycle: move all database tables to the entities folder for easy tracking and migrations
