+++ src/cross-platform/services/scheduler.service.spec.ts
@@ -5,7 +5,7 @@ import { BadRequestException, Logger, NotFoundException } from '@nestjs/common';
 import { SchedulerService } from './scheduler.service';
 import { ContentPublisherService } from './content-publisher.service';
 import { MediaStorageService } from '../../asset-management/media-storage/media-storage.service';
-import { ScheduledPost } from '../entity/schedule.entity';
+import { ScheduledPost } from '../../entities/cross-platform-entities/schedule.entity';
 import { SocialPlatform } from '../../common/enums/social-platform.enum';
 import {
   PublishStatus,
@@ -131,7 +131,7 @@ describe('SchedulerService', () => {
     // Mock Logger
     loggerSpy = jest
       .spyOn(Logger.prototype, 'error')
-      .mockImplementation(() => {});
+      .mockImplementation(() => { });
   });
 
   afterEach(() => {
