+++ src/cross-platform/services/scheduler.service.ts
@@ -12,7 +12,7 @@ import {
   ScheduleStatus,
 } from '../helpers/cross-platform.interface';
 import { ContentPublisherService } from './content-publisher.service';
-import { ScheduledPost } from '../entity/schedule.entity';
+import { ScheduledPost } from '../../entities/cross-platform-entities/schedule.entity';
 import { SocialPlatform } from '../../common/enums/social-platform.enum';
 import { MediaStorageService } from '../../asset-management/media-storage/media-storage.service';
 import { MediaStorageItem } from '../../asset-management/media-storage/media-storage.dto';
@@ -25,7 +25,7 @@ export class SchedulerService {
     private readonly scheduledPostRepo: Repository<ScheduledPost>,
     private readonly contentPublisherService: ContentPublisherService,
     private readonly mediaStorageService: MediaStorageService,
-  ) {}
+  ) { }
 
   @Cron(CronExpression.EVERY_MINUTE)
   async processScheduledPosts() {
