+++ src/entities/notifications/entity/social-account.entity.ts
@@ -6,9 +6,11 @@ import {
   UpdateDateColumn,
   OneToOne,
 } from 'typeorm';
-import { FacebookAccount } from '../facebook/entity/facebook-account.entity';
+
 import { TenantEntity } from './tenant-entity';
-import { SocialPlatform } from '../../common/enums/social-platform.enum';
+import { SocialPlatform } from '../../../common/enums/social-platform.enum';
+import { FacebookAccount } from '../../socials/facebook-entities/facebook-account.entity';
+
 
 @Entity('social_accounts')
 export class SocialAccount extends TenantEntity {
