+++ src/entities/socials/facebook-entities/auth-state.entity.ts
