+++ src/entities/socials/facebook-entities/facebook-account.entity.ts
@@ -9,8 +9,8 @@ import {
   JoinColumn,
 } from 'typeorm';
 import { FacebookPage } from './facebook-page.entity';
-import { SocialAccount } from '../../entity/social-account.entity';
-import { TenantEntity } from '../../entity/tenant-entity';
+import { SocialAccount } from '../../notifications/entity/social-account.entity';
+import { TenantEntity } from '../../notifications/entity/tenant-entity';
 
 @Entity('facebook_accounts')
 export class FacebookAccount extends TenantEntity {
