+++ src/entities/socials/facebook-entities/facebook-page.entity.ts
@@ -9,7 +9,8 @@ import {
 } from 'typeorm';
 import { FacebookAccount } from './facebook-account.entity';
 import { FacebookPost } from './facebook-post.entity';
-import { TenantEntity } from '../../entity/tenant-entity';
+import { TenantEntity } from '../../notifications/entity/tenant-entity';
+
 
 @Entity('facebook_pages')
 export class FacebookPage extends TenantEntity {
