+++ src/entities/socials/instagram-entities/instagram-metric.entity.ts
@@ -6,7 +6,8 @@ import {
   ManyToOne,
 } from 'typeorm';
 import { InstagramPost } from './instagram-post.entity';
-import { TenantEntity } from '../../entity/tenant-entity';
+import { TenantEntity } from '../../notifications/entity/tenant-entity';
+
 
 @Entity('instagram_metrics')
 export class InstagramMetric extends TenantEntity {
