+++ src/entities/socials/linkedin-entities/linkedin-organization.entity.ts
@@ -9,7 +9,8 @@ import {
 } from 'typeorm';
 import { LinkedInAccount } from './linkedin-account.entity';
 import { LinkedInPost } from './linkedin-post.entity';
-import { TenantEntity } from '../../../platforms/entity/tenant-entity';
+import { TenantEntity } from '../../notifications/entity/tenant-entity';
+
 
 @Entity('linkedin_organizations')
 export class LinkedInOrganization extends TenantEntity {
