+++ src/entities/socials/tiktok-entities/tiktok-video.entity.ts
@@ -9,8 +9,8 @@ import {
 } from 'typeorm';
 import { TikTokAccount } from './tiktok-account.entity';
 import { TikTokMetric } from './tiktok-metric.entity';
-import { TikTokVideoPrivacyLevel } from '../helpers/tiktok.interfaces';
-import { TenantEntity } from '../../../platforms/entity/tenant-entity';
+import { TikTokVideoPrivacyLevel } from '../../../platforms/tiktok/helpers/tiktok.interfaces';
+import { TenantEntity } from '../../notifications/entity/tenant-entity';
 
 @Entity('tiktok_videos')
 export class TikTokVideo extends TenantEntity {
