+++ src/entities/socials/tiktok-entities/tiktok_comments.entity.ts
@@ -7,7 +7,8 @@ import {
   ManyToOne,
 } from 'typeorm';
 import { TikTokVideo } from './tiktok-video.entity';
-import { TenantEntity } from '../../../platforms/entity/tenant-entity';
+import { TenantEntity } from '../../notifications/entity/tenant-entity';
+
 
 @Entity('tiktok_comments')
 export class TikTokComment extends TenantEntity {
