+++ src/platform-auth/platform-auth.module.ts
@@ -7,30 +7,30 @@ import { FacebookRepository } from '../platforms/facebook/repositories/facebook.
 import { InstagramRepository } from '../platforms/instagram/repositories/instagram.repository';
 import { LinkedInRepository } from '../platforms/linkedin/repositories/linkedin.repository';
 import { TikTokRepository } from '../platforms/tiktok/repositories/tiktok.repository';
-import { FacebookAccount } from '../platforms/facebook/entity/facebook-account.entity';
-import { SocialAccount } from '../platforms/entity/social-account.entity';
-import { AuthState } from '../platforms/facebook/entity/auth-state.entity';
-import { FacebookPageMetric } from '../platforms/facebook/entity/facebook-page-metric.entity';
-import { FacebookPage } from '../platforms/facebook/entity/facebook-page.entity';
-import { FacebookPostMetric } from '../platforms/facebook/entity/facebook-post-metric.entity';
-import { FacebookPost } from '../platforms/facebook/entity/facebook-post.entity';
-import { InstagramAccount } from '../platforms/instagram/entities/instagram-account.entity';
-import { InstagramMetric } from '../platforms/instagram/entities/instagram-metric.entity';
-import { InstagramPost } from '../platforms/instagram/entities/instagram-post.entity';
-import { InstagramRateLimit } from '../platforms/instagram/entities/instagram-rate-limit.entity';
-import { LinkedInAccount } from '../platforms/linkedin/entities/linkedin-account.entity';
-import { LinkedInOrganization } from '../platforms/linkedin/entities/linkedin-organization.entity';
+import { LinkedInAccount } from '../entities/socials/linkedin-entities/linkedin-account.entity';
+import { LinkedInOrganization } from '../entities/socials/linkedin-entities/linkedin-organization.entity';
 import {
   LinkedInPost,
   LinkedInMetric,
-} from '../platforms/linkedin/entities/linkedin-post.entity';
-import { TikTokAccount } from '../platforms/tiktok/entities/tiktok-account.entity';
-import { TikTokMetric } from '../platforms/tiktok/entities/tiktok-metric.entity';
-import { TikTokVideo } from '../platforms/tiktok/entities/tiktok-video.entity';
-import { TikTokComment } from '../platforms/tiktok/entities/tiktok_comments.entity';
-import { TikTokRateLimit } from '../platforms/tiktok/entities/tiktok_rate_limits.entity';
+} from '../entities/socials/linkedin-entities/linkedin-post.entity';
+import { TikTokAccount } from '../entities/socials/tiktok-entities/tiktok-account.entity';
+import { TikTokMetric } from '../entities/socials/tiktok-entities/tiktok-metric.entity';
+import { TikTokVideo } from '../entities/socials/tiktok-entities/tiktok-video.entity';
+import { TikTokComment } from '../entities/socials/tiktok-entities/tiktok_comments.entity';
+import { TikTokRateLimit } from '../entities/socials/tiktok-entities/tiktok_rate_limits.entity';
 import { CacheModule } from '../cache/cache.module';
 import { SocialPlatform } from '../common/enums/social-platform.enum';
+import { SocialAccount } from '../entities/notifications/entity/social-account.entity';
+import { AuthState } from '../entities/socials/facebook-entities/auth-state.entity';
+import { FacebookAccount } from '../entities/socials/facebook-entities/facebook-account.entity';
+import { FacebookPageMetric } from '../entities/socials/facebook-entities/facebook-page-metric.entity';
+import { FacebookPage } from '../entities/socials/facebook-entities/facebook-page.entity';
+import { FacebookPostMetric } from '../entities/socials/facebook-entities/facebook-post-metric.entity';
+import { FacebookPost } from '../entities/socials/facebook-entities/facebook-post.entity';
+import { InstagramAccount } from '../entities/socials/instagram-entities/instagram-account.entity';
+import { InstagramMetric } from '../entities/socials/instagram-entities/instagram-metric.entity';
+import { InstagramPost } from '../entities/socials/instagram-entities/instagram-post.entity';
+import { InstagramRateLimit } from '../entities/socials/instagram-entities/instagram-rate-limit.entity';
 
 @Module({
   providers: [
@@ -106,4 +106,4 @@ import { SocialPlatform } from '../common/enums/social-platform.enum';
     ]),
   ],
 })
-export class PlatformAuthModule {}
+export class PlatformAuthModule { }
