+++ src/platforms/facebook/facebook.module.ts
@@ -6,16 +6,16 @@ import { FacebookPageInsightsJob } from './jobs/page-insight.job';
 import { FacebookService } from './facebook.service';
 import { FacebookPostMetricsJob } from './jobs/post-metrics-collection.job';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FacebookAccount } from './entity/facebook-account.entity';
-import { FacebookPage } from './entity/facebook-page.entity';
-import { FacebookPost } from './entity/facebook-post.entity';
-import { FacebookPostMetric } from './entity/facebook-post-metric.entity';
-import { FacebookPageMetric } from './entity/facebook-page-metric.entity';
-import { AuthState } from './entity/auth-state.entity';
-import { SocialAccount } from '../entity/social-account.entity';
 import { MediaStorageModule } from '../../asset-management/media-storage/media-storage.module';
 import { DatabaseModule } from '../../database/database.module';
 import { PlatformAuthModule } from '../../platform-auth/platform-auth.module';
+import { SocialAccount } from '../../entities/notifications/entity/social-account.entity';
+import { AuthState } from '../../entities/socials/facebook-entities/auth-state.entity';
+import { FacebookAccount } from '../../entities/socials/facebook-entities/facebook-account.entity';
+import { FacebookPageMetric } from '../../entities/socials/facebook-entities/facebook-page-metric.entity';
+import { FacebookPage } from '../../entities/socials/facebook-entities/facebook-page.entity';
+import { FacebookPostMetric } from '../../entities/socials/facebook-entities/facebook-post-metric.entity';
+import { FacebookPost } from '../../entities/socials/facebook-entities/facebook-post.entity';
 
 @Module({
   controllers: [FacebookController],
