+++ src/platforms/instagram/repositories/instagram.repository.ts
@@ -2,14 +2,18 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectEntityManager, InjectRepository } from '@nestjs/typeorm';
 import { Repository, MoreThan, LessThan, EntityManager } from 'typeorm';
 import * as crypto from 'crypto';
-import { InstagramAccount } from '../entities/instagram-account.entity';
-import { InstagramMetric } from '../entities/instagram-metric.entity';
-import { InstagramRateLimit } from '../entities/instagram-rate-limit.entity';
-import { InstagramPost } from '../entities/instagram-post.entity';
+
 import { SocialPlatform } from '../../../common/enums/social-platform.enum';
-import { SocialAccount } from '../../entity/social-account.entity';
-import { AuthState } from '../../facebook/entity/auth-state.entity';
+
 import { TenantAwareRepository } from '../../../database/tenant-aware.repository';
+import { SocialAccount } from '../../../entities/notifications/entity/social-account.entity';
+import { AuthState } from '../../../entities/socials/facebook-entities/auth-state.entity';
+import { InstagramMetric } from '../../../entities/socials/instagram-entities/instagram-metric.entity';
+import { InstagramPost } from '../../../entities/socials/instagram-entities/instagram-post.entity';
+import { InstagramRateLimit } from '../../../entities/socials/instagram-entities/instagram-rate-limit.entity';
+import { InstagramAccount } from '../../../entities/socials/instagram-entities/instagram-account.entity';
+
+
 
 @Injectable()
 export class InstagramRepository extends TenantAwareRepository {
