+++ src/platforms/linkedin/linkedin.service.ts
@@ -22,7 +22,7 @@ import {
   DateRange,
   PostMetrics,
 } from '../../cross-platform/helpers/cross-platform.interface';
-import { LinkedInAccount } from './entities/linkedin-account.entity';
+import { LinkedInAccount } from '../../entities/socials/linkedin-entities/linkedin-account.entity';
 import { TenantService } from '../../database/tenant.service';
 import { CreateLinkedInPostDto } from './helpers/create-post.dto';
 import { MediaStorageService } from '../../asset-management/media-storage/media-storage.service';
@@ -38,7 +38,7 @@ export class LinkedInService implements PlatformService {
     private readonly linkedInRepo: LinkedInRepository,
     private readonly tenantService: TenantService,
     private readonly mediaStorageService: MediaStorageService,
-  ) {}
+  ) { }
 
   private async uploadLinkedInMediaItemsToStorage(
     media: MediaItem[],
