+++ src/platforms/linkedin/repositories/linkedin.repository.ts
@@ -2,14 +2,14 @@ import * as crypto from 'crypto';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectEntityManager, InjectRepository } from '@nestjs/typeorm';
 import { EntityManager, LessThan, MoreThan, Repository } from 'typeorm';
-import { LinkedInAccount } from '../entities/linkedin-account.entity';
-import { LinkedInOrganization } from '../entities/linkedin-organization.entity';
-import { LinkedInMetric } from '../entities/linkedin-metric.entity';
-import { SocialAccount } from '../../../platforms/entity/social-account.entity';
-import { LinkedInPost } from '../entities/linkedin-post.entity';
-import { AuthState } from '../../../platforms/facebook/entity/auth-state.entity';
+import { LinkedInAccount } from '../../../entities/socials/linkedin-entities/linkedin-account.entity';
+import { LinkedInOrganization } from '../../../entities/socials/linkedin-entities/linkedin-organization.entity';
+import { LinkedInMetric } from '../../../entities/socials/linkedin-entities/linkedin-metric.entity';
+import { LinkedInPost } from '../../../entities/socials/linkedin-entities/linkedin-post.entity';
 import { TenantAwareRepository } from '../../../database/tenant-aware.repository';
 import { SocialPlatform } from '../../../common/enums/social-platform.enum';
+import { SocialAccount } from '../../../entities/notifications/entity/social-account.entity';
+import { AuthState } from '../../../entities/socials/facebook-entities/auth-state.entity';
 
 @Injectable()
 export class LinkedInRepository extends TenantAwareRepository {
