+++ src/platforms/tiktok/tiktok.service.ts
@@ -28,7 +28,7 @@ import {
   DateRange,
   PostMetrics,
 } from '../../cross-platform/helpers/cross-platform.interface';
-import { TikTokAccount } from './entities/tiktok-account.entity';
+import { TikTokAccount } from '../../entities/socials/tiktok-entities/tiktok-account.entity';
 import { TenantService } from '../../database/tenant.service';
 import { MediaStorageService } from '../../asset-management/media-storage/media-storage.service';
 import { MediaStorageItem } from '../../asset-management/media-storage/media-storage.dto';
