+++ src/platforms/tiktok/titkot.module.ts
@@ -12,21 +12,21 @@ import { TikTokConfig } from './config/tiktok.config';
 // Shared imports
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { SocialAccount } from '../entity/social-account.entity';
-import { AuthState } from '../facebook/entity/auth-state.entity';
 
 // TikTok entities
-import { TikTokAccount } from './entities/tiktok-account.entity';
-import { TikTokVideo } from './entities/tiktok-video.entity';
-import { TikTokMetric } from './entities/tiktok-metric.entity';
-import { TikTokComment } from './entities/tiktok_comments.entity';
-import { TikTokRateLimit } from './entities/tiktok_rate_limits.entity';
+import { TikTokAccount } from '../../entities/socials/tiktok-entities/tiktok-account.entity';
+import { TikTokVideo } from '../../entities/socials/tiktok-entities/tiktok-video.entity';
+import { TikTokMetric } from '../../entities/socials/tiktok-entities/tiktok-metric.entity';
+import { TikTokComment } from '../../entities/socials/tiktok-entities/tiktok_comments.entity';
+import { TikTokRateLimit } from '../../entities/socials/tiktok-entities/tiktok_rate_limits.entity';
 
 // Repositories for other platforms
 import { TikTokTokenRefreshJob } from './jobs/token-refresh.job';
 import { DatabaseModule } from '../../database/database.module';
 import { PlatformAuthModule } from '../../platform-auth/platform-auth.module';
 import { MediaStorageModule } from '../../asset-management/media-storage/media-storage.module';
+import { SocialAccount } from '../../entities/notifications/entity/social-account.entity';
+import { AuthState } from '../../entities/socials/facebook-entities/auth-state.entity';
 
 @Module({
   controllers: [TikTokController],
@@ -60,4 +60,4 @@ import { MediaStorageModule } from '../../asset-management/media-storage/media-s
   ],
   exports: [TikTokService],
 })
-export class TikTokModule {}
+export class TikTokModule { }
